+++ spec/httpi/adapter/em_http_spec.rb
@@ -150,7 +150,7 @@ begin
 
         it "raises an error for HTTP digest auth" do
           request.auth.digest "username", "password"
-          expect { adapter.request(:get) }.to raise_error
+          expect { adapter.request(:get) }.to raise_error HTTPI::NotSupportedError
         end
       end